renamed field_choice.py to dynamic_fields.py
